pyproject.toml
[project]
name = "class-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.5",
    "pillow-avif-plugin>=1.5.2",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "scikit-learn>=1.7.1",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "urllib3>=2.5.0",
]
